chore: extract unroute behavior tests into a separate file
Reference https://github.com/microsoft/playwright/issues/23781
